EdgeSolution-PSUtil-Demo-01/modules/module_Controller/requirements-module.txt
azure-iot-device==2.7.1
psutil==5.9.4